change int types to bigger ints for better robustness

+++ src/clique.cpp
@@ -27,10 +27,10 @@ bool Clique::hasVertex(Vertex *v) const
     return false;
 }
 
-int Clique::computeWeight() const
+long unsigned int Clique::computeWeight() const
 {
     // if an edge exists between two vertices in the clique, add its weight to the total
-    int weight = 0;
+    long unsigned int weight = 0;
     for (auto &vertex : vertices)
         for (auto &e : vertex->getEdges())
             if (hasVertex(e->getV1() == vertex ? e->getV2() : e->getV1()))
@@ -40,22 +40,22 @@ int Clique::computeWeight() const
     return weight / 2;
 }
 
-void Clique::setWeight(int weight)
+void Clique::setWeight(long unsigned int weight)
 {
     this->weight = weight;
 }
 
-void Clique::addWeight(int weight)
+void Clique::addWeight(long unsigned int weight)
 {
     this->weight += weight;
 }
 
-int Clique::getWeight() const
+long unsigned int Clique::getWeight() const
 {
     return weight;
 }
 
-int Clique::getSize() const
+long unsigned int Clique::getSize() const
 {
     return vertices.size();
 }
